requirements.txt
aliyun-python-sdk-core-v3==2.12.1
aliyun-python-sdk-kms==2.5.0
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.5
Django==1.11.7
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-xadmin==0.5.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
future==0.17.1
gunicorn==19.9.0
httplib2==0.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jpush==3.3.5
Markdown==3.0.1
MarkupSafe==1.1.0
oss2==2.6.1
Pillow==5.4.1
pycparser==2.19
pycryptodome==3.7.3
PyJWT==1.7.1
pytz==2018.7
PyYAML==3.13
raven==6.10.0
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
xadmin==0.6.1
